// src/components/RoundButton/RoundButton.types.ts
export enum ButtonSizes {
  SMALL = 'size-[min(13.5vh,13.5vw)]', // 111px
  MEDIUM = 'size-[min(19vh,19vw)]', // 155px
  LARGE = 'size-[min(27vh,27vw)]', // 225px
}

export enum BorderSizes {
  SMALL = 'border-[min(2vh,2vw)]', // 17px
  MEDIUM = 'border-[min(2.5vh,2.5vw)]', // 24px
  LARGE = 'border-[min(3.5vh,3.5vw)]', // 34px
}

export enum ImageSizes {
  SMALL = 'w-[min(5vh,5vw)]',
  MEDIUM = 'w-[min(7vh,7vw)]',
  LARGE = 'w-[min(11vh,11vw)]',
}

export enum LongTurnsList {
  PAPER = 'paper',
  ROCK = 'rock',
  SCISSORS = 'scissors',
  LIZARD = 'lizard',
  SPOCK = 'spock',
}

export enum ShortTurnsList {
  PAPER = 'paper',
  ROCK = 'rock',
  SCISSORS = 'scissors',
}